Split PageSerde into PageSerializer and PageDeserializer
Usually a single call site (such as an operator) does not need both.
Hence there is a logic for lazy buffer allocation for serialize and
deserialize. Lazy allocation makes memory accounting less
straightforward.
