+++ core/trino-main/src/main/java/io/trino/operator/MergeOperator.java
@@ -17,7 +17,7 @@ import com.google.common.io.Closer;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 import io.trino.exchange.DirectExchangeInput;
-import io.trino.execution.buffer.PagesSerde;
+import io.trino.execution.buffer.PageDeserializer;
 import io.trino.execution.buffer.PagesSerdeFactory;
 import io.trino.metadata.Split;
 import io.trino.spi.Page;
@@ -102,7 +102,7 @@ public class MergeOperator
                     operatorContext,
                     sourceId,
                     directExchangeClientSupplier,
-                    serdeFactory.createPagesSerde(driverContext.getSession().getExchangeEncryptionKey().map(Ciphers::deserializeAesEncryptionKey)),
+                    serdeFactory.createDeserializer(driverContext.getSession().getExchangeEncryptionKey().map(Ciphers::deserializeAesEncryptionKey)),
                     orderingCompiler.compilePageWithPositionComparator(types, sortChannels, sortOrder),
                     outputChannels,
                     outputTypes);
@@ -118,7 +118,7 @@ public class MergeOperator
     private final OperatorContext operatorContext;
     private final PlanNodeId sourceId;
     private final DirectExchangeClientSupplier directExchangeClientSupplier;
-    private final PagesSerde pagesSerde;
+    private final PageDeserializer deserializer;
     private final PageWithPositionComparator comparator;
     private final List<Integer> outputChannels;
     private final List<Type> outputTypes;
@@ -135,7 +135,7 @@ public class MergeOperator
             OperatorContext operatorContext,
             PlanNodeId sourceId,
             DirectExchangeClientSupplier directExchangeClientSupplier,
-            PagesSerde pagesSerde,
+            PageDeserializer deserializer,
             PageWithPositionComparator comparator,
             List<Integer> outputChannels,
             List<Type> outputTypes)
@@ -143,7 +143,7 @@ public class MergeOperator
         this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
         this.sourceId = requireNonNull(sourceId, "sourceId is null");
         this.directExchangeClientSupplier = requireNonNull(directExchangeClientSupplier, "directExchangeClientSupplier is null");
-        this.pagesSerde = requireNonNull(pagesSerde, "pagesSerde is null");
+        this.deserializer = requireNonNull(deserializer, "deserializer is null");
         this.comparator = requireNonNull(comparator, "comparator is null");
         this.outputChannels = requireNonNull(outputChannels, "outputChannels is null");
         this.outputTypes = requireNonNull(outputTypes, "outputTypes is null");
@@ -177,7 +177,7 @@ public class MergeOperator
         client.noMoreLocations();
         pageProducers.add(client.pages()
                 .map(serializedPage -> {
-                    Page page = pagesSerde.deserialize(serializedPage);
+                    Page page = deserializer.deserialize(serializedPage);
                     operatorContext.recordNetworkInput(serializedPage.length(), page.getPositionCount());
                     return page;
                 }));
