+++ core/trino-main/src/test/java/io/trino/execution/buffer/BenchmarkBlockSerde.java
@@ -92,7 +92,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeLongDecimal(LongDecimalBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -104,7 +104,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeInt96(LongTimestampBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -116,7 +116,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeLong(BigintBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -128,7 +128,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeInteger(IntegerBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -140,7 +140,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeShort(SmallintBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -152,7 +152,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeByte(TinyintBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -164,7 +164,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeSliceDirect(VarcharDirectBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -176,7 +176,7 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeLineitem(LineitemBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     @Benchmark
@@ -188,13 +188,13 @@ public class BenchmarkBlockSerde
     @Benchmark
     public Object deserializeRow(RowTypeBenchmarkData data)
     {
-        return ImmutableList.copyOf(readPages(data.getPagesSerde(), new BasicSliceInput(data.getDataSource())));
+        return ImmutableList.copyOf(readPages(data.getDeserializer(), new BasicSliceInput(data.getDataSource())));
     }
 
     private static List<Slice> serializePages(BenchmarkData data)
     {
         return data.getPages().stream()
-                .map(page -> data.getPagesSerde().serialize(page))
+                .map(page -> data.getSerializer().serialize(page))
                 .collect(toImmutableList());
     }
 
@@ -206,7 +206,9 @@ public class BenchmarkBlockSerde
 
         public void setup(Type type, Function<Random, ?> valueGenerator)
         {
-            PagesSerde pagesSerde = new PagesSerdeFactory(new TestingBlockEncodingSerde(), false).createPagesSerde(Optional.empty());
+            PagesSerdeFactory serdeFactory = new PagesSerdeFactory(new TestingBlockEncodingSerde(), false);
+            PageSerializer serializer = serdeFactory.createSerializer(Optional.empty());
+            PageDeserializer deserializer = serdeFactory.createDeserializer(Optional.empty());
             PageBuilder pageBuilder = new PageBuilder(ImmutableList.of(type));
             BlockBuilder blockBuilder = pageBuilder.getBlockBuilder(0);
             ImmutableList.Builder<Page> pagesBuilder = ImmutableList.builder();
@@ -227,9 +229,9 @@ public class BenchmarkBlockSerde
 
             List<Page> pages = pagesBuilder.build();
             DynamicSliceOutput sliceOutput = new DynamicSliceOutput(0);
-            writePages(pagesSerde, new OutputStreamSliceOutput(sliceOutput), pages.iterator());
+            writePages(serializer, new OutputStreamSliceOutput(sliceOutput), pages.iterator());
 
-            setup(sliceOutput.slice(), pagesSerde, pages);
+            setup(sliceOutput.slice(), serializer, deserializer, pages);
         }
 
         private void writeValue(Type type, Object value, BlockBuilder blockBuilder)
@@ -281,13 +283,15 @@ public class BenchmarkBlockSerde
     public abstract static class BenchmarkData
     {
         private Slice dataSource;
-        private PagesSerde pagesSerde;
+        private PageSerializer serializer;
+        private PageDeserializer deserializer;
         private List<Page> pages;
 
-        public void setup(Slice dataSource, PagesSerde pagesSerde, List<Page> pages)
+        public void setup(Slice dataSource, PageSerializer serializer, PageDeserializer deserializer, List<Page> pages)
         {
             this.dataSource = dataSource;
-            this.pagesSerde = pagesSerde;
+            this.serializer = serializer;
+            this.deserializer = deserializer;
             this.pages = pages;
         }
 
@@ -296,9 +300,14 @@ public class BenchmarkBlockSerde
             return pages;
         }
 
-        public PagesSerde getPagesSerde()
+        public PageSerializer getSerializer()
         {
-            return pagesSerde;
+            return serializer;
+        }
+
+        public PageDeserializer getDeserializer()
+        {
+            return deserializer;
         }
 
         public Slice getDataSource()
@@ -391,12 +400,14 @@ public class BenchmarkBlockSerde
         @Setup
         public void setup()
         {
-            PagesSerde pagesSerde = new PagesSerdeFactory(new TestingBlockEncodingSerde(), false).createPagesSerde(Optional.empty());
+            PagesSerdeFactory serdeFactory = new PagesSerdeFactory(new TestingBlockEncodingSerde(), false);
+            PageSerializer serializer = serdeFactory.createSerializer(Optional.empty());
+            PageDeserializer deserializer = serdeFactory.createDeserializer(Optional.empty());
 
             List<Page> pages = ImmutableList.copyOf(getTablePages("lineitem", 0.1, DecimalTypeMapping.DOUBLE));
             DynamicSliceOutput sliceOutput = new DynamicSliceOutput(0);
-            writePages(pagesSerde, new OutputStreamSliceOutput(sliceOutput), pages.listIterator());
-            setup(sliceOutput.slice(), pagesSerde, pages);
+            writePages(serializer, new OutputStreamSliceOutput(sliceOutput), pages.listIterator());
+            setup(sliceOutput.slice(), serializer, deserializer, pages);
         }
     }
