+++ core/trino-main/src/test/java/io/trino/operator/TestHttpPageBufferClient.java
@@ -27,7 +27,9 @@ import io.airlift.units.Duration;
 import io.trino.FeaturesConfig.DataIntegrityVerification;
 import io.trino.execution.StageId;
 import io.trino.execution.TaskId;
-import io.trino.execution.buffer.PagesSerde;
+import io.trino.execution.buffer.PageDeserializer;
+import io.trino.execution.buffer.PagesSerdeFactory;
+import io.trino.execution.buffer.TestingPagesSerdeFactory;
 import io.trino.operator.HttpPageBufferClient.ClientCallback;
 import io.trino.spi.HostAddress;
 import io.trino.spi.Page;
@@ -41,6 +43,7 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.ExecutorService;
@@ -59,7 +62,6 @@ import static io.airlift.testing.Assertions.assertContains;
 import static io.airlift.testing.Assertions.assertInstanceOf;
 import static io.airlift.units.DataSize.Unit.MEGABYTE;
 import static io.trino.TrinoMediaTypes.TRINO_PAGES;
-import static io.trino.execution.buffer.TestingPagesSerdeFactory.testingPagesSerde;
 import static io.trino.spi.StandardErrorCode.EXCEEDED_LOCAL_MEMORY_LIMIT;
 import static io.trino.spi.StandardErrorCode.PAGE_TOO_LARGE;
 import static io.trino.spi.StandardErrorCode.PAGE_TRANSPORT_ERROR;
@@ -75,7 +77,6 @@ public class TestHttpPageBufferClient
     private ScheduledExecutorService scheduler;
     private ExecutorService pageBufferClientCallbackExecutor;
 
-    private static final PagesSerde PAGES_SERDE = testingPagesSerde();
     private static final TaskId TASK_ID = new TaskId(new StageId("query", 0), 0, 0);
 
     @BeforeClass
@@ -516,6 +517,8 @@ public class TestHttpPageBufferClient
     private static class TestingClientCallback
             implements ClientCallback
     {
+        private final PagesSerdeFactory serdeFactory = new TestingPagesSerdeFactory();
+
         private final CyclicBarrier done;
         private final List<Slice> pages = Collections.synchronizedList(new ArrayList<>());
         private final AtomicInteger completedRequests = new AtomicInteger();
@@ -530,8 +533,9 @@ public class TestHttpPageBufferClient
 
         public List<Page> getPages()
         {
+            PageDeserializer deserializer = serdeFactory.createDeserializer(Optional.empty());
             return pages.stream()
-                    .map(PAGES_SERDE::deserialize)
+                    .map(deserializer::deserialize)
                     .collect(Collectors.toList());
         }
