+++ core/trino-main/src/test/java/io/trino/spiller/TestFileSingleStreamSpiller.java
@@ -39,8 +39,8 @@ import static com.google.common.io.MoreFiles.deleteRecursively;
 import static com.google.common.io.MoreFiles.listFiles;
 import static com.google.common.io.RecursiveDeleteOption.ALLOW_INSECURE;
 import static com.google.common.util.concurrent.MoreExecutors.listeningDecorator;
-import static io.trino.execution.buffer.PagesSerde.isSerializedPageCompressed;
-import static io.trino.execution.buffer.PagesSerde.isSerializedPageEncrypted;
+import static io.trino.execution.buffer.PagesSerdeUtil.isSerializedPageCompressed;
+import static io.trino.execution.buffer.PagesSerdeUtil.isSerializedPageEncrypted;
 import static io.trino.memory.context.AggregatedMemoryContext.newSimpleAggregatedMemoryContext;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.DoubleType.DOUBLE;
